scripts+tapdb: do sqlc required replacement in script
Our SQL scripts on disk were neither SQLite nor Postgres compatible, we
used a version suited for sqlc to generate the correct data types we
wanted, then did some type replacements in-memory when actually applying
the migration files.
This is somewhat dangerous as it will lead to situations where if a user
applies a migration file manually, they'll end up with table primary
keys that ARE NOT auto incrementing.

To fix this issue, we make sure that we always write fully SQLite
compatible SQL code and only apply fixes for the sqlc code generation
directly before generating the code.

+++ tapdb/sqlc/migrations/000005_transfers.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS asset_transfers (
-    id BIGINT PRIMARY KEY, 
+    id INTEGER PRIMARY KEY, 
 
     height_hint INTEGER NOT NULL,
     
@@ -13,7 +13,7 @@ CREATE INDEX IF NOT EXISTS transfer_txn_idx
     ON asset_transfers (anchor_txn_id);
 
 CREATE TABLE IF NOT EXISTS asset_transfer_inputs (
-    input_id BIGINT PRIMARY KEY,
+    input_id INTEGER PRIMARY KEY,
     
     transfer_id BIGINT NOT NULL REFERENCES asset_transfers(id),
     
@@ -29,7 +29,7 @@ CREATE INDEX IF NOT EXISTS transfer_inputs_idx
     ON asset_transfer_inputs (transfer_id);
 
 CREATE TABLE IF NOT EXISTS asset_transfer_outputs (
-    output_id BIGINT PRIMARY KEY,
+    output_id INTEGER PRIMARY KEY,
     
     transfer_id BIGINT NOT NULL REFERENCES asset_transfers(id),
     
@@ -74,7 +74,7 @@ CREATE INDEX IF NOT EXISTS proof_locator_hash_index
 -- passive_assets is a table that stores the information needed to
 -- re-anchor a passive asset.
 CREATE TABLE IF NOT EXISTS passive_assets (
-    passive_id BIGINT PRIMARY KEY,
+    passive_id INTEGER PRIMARY KEY,
 
     transfer_id BIGINT NOT NULL REFERENCES asset_transfers(id),
