+++ tapdb/sqlc/migrations/000007_universe.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS universe_roots (
-    id BIGINT PRIMARY KEY,
+    id INTEGER PRIMARY KEY,
 
     -- For the namespace root, we set the foreign key constraint evaluation to
     -- be deferred until after the database transaction ends. Otherwise, if the
@@ -23,7 +23,7 @@ CREATE INDEX IF NOT EXISTS universe_roots_asset_id_idx ON universe_roots(asset_i
 CREATE INDEX IF NOT EXISTS universe_roots_group_key_idx ON universe_roots(group_key);
 
 CREATE TABLE IF NOT EXISTS universe_leaves (
-    id BIGINT PRIMARY KEY,
+    id INTEGER PRIMARY KEY,
 
     asset_genesis_id BIGINT NOT NULL REFERENCES genesis_assets(gen_asset_id),
 
@@ -44,7 +44,7 @@ CREATE INDEX IF NOT EXISTS universe_leaves_key_idx ON universe_leaves(leaf_node_
 CREATE INDEX IF NOT EXISTS universe_leaves_namespace ON universe_leaves(leaf_node_namespace);
 
 CREATE TABLE IF NOT EXISTS universe_servers (
-    id BIGINT PRIMARY KEY,
+    id INTEGER PRIMARY KEY,
 
     server_host TEXT UNIQUE NOT NULL,
 
@@ -59,7 +59,7 @@ CREATE TABLE IF NOT EXISTS universe_servers (
 CREATE INDEX IF NOT EXISTS universe_servers_host ON universe_servers(server_host);
 
 CREATE TABLE IF NOT EXISTS universe_events (
-    event_id BIGINT PRIMARY KEY,
+    event_id INTEGER PRIMARY KEY,
 
     event_type VARCHAR NOT NULL CHECK (event_type IN ('SYNC', 'NEW_PROOF', 'NEW_ROOT')),
